break out amount select
